BasicFragment + rework others Fragments
