+++ app/src/main/java/com/example/myassets/presentation/ui/assets/AssetListFragment.kt
@@ -1,27 +1,13 @@
 package com.example.myassets.presentation.ui.assets
 
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import com.example.myassets.databinding.FragmentAssetListBinding
+import com.example.myassets.presentation.util.BasicFragment
 
-class AssetListFragment : Fragment() {
-    private var _binding: FragmentAssetListBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
+class AssetListFragment : BasicFragment<FragmentAssetListBinding>() {
+    override fun inflateViewBinding(
         inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentAssetListBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+        container: ViewGroup?
+    ) = FragmentAssetListBinding.inflate(inflater, container, false)
 }
