+++ app/src/main/java/com/example/myassets/presentation/ui/home/HomeFragment.kt
@@ -1,27 +1,13 @@
 package com.example.myassets.presentation.ui.home
 
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import com.example.myassets.databinding.FragmentHomeBinding
+import com.example.myassets.presentation.util.BasicFragment
 
-class HomeFragment : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
+class HomeFragment : BasicFragment<FragmentHomeBinding>() {
+    override fun inflateViewBinding(
         inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+        container: ViewGroup?
+    ) = FragmentHomeBinding.inflate(inflater, container, false)
 }
