+++ app/src/main/java/com/example/myassets/presentation/ui/settings/SettingsFragment.kt
@@ -1,27 +1,13 @@
 package com.example.myassets.presentation.ui.settings
 
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import com.example.myassets.databinding.FragmentSettingsBinding
+import com.example.myassets.presentation.util.BasicFragment
 
-class SettingsFragment : Fragment() {
-    private var _binding: FragmentSettingsBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
+class SettingsFragment : BasicFragment<FragmentSettingsBinding>() {
+    override fun inflateViewBinding(
         inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentSettingsBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+        container: ViewGroup?
+    ) = FragmentSettingsBinding.inflate(inflater, container, false)
 }
